second commit

+++ Graph.cpp
@@ -2,12 +2,12 @@
 
 using namespace ariel;
 
-void Graph::loadGraph(const std::vector<std::vector<int>>& graph) {
+void Graph::loadGraph(const std::vector<std::vector<size_t>>& graph) {
     matrix = graph;
 }
 
 void Graph::printGraph() const {
-    int vertices = 0;
+    size_t vertices = 0;
 
     for (const auto& innerVec : matrix) {
         for (int num : innerVec) {
@@ -21,10 +21,10 @@ void Graph::printGraph() const {
 }
 
 // Help Functions
-int Graph::size() const{
+size_t Graph::size() const{
     return matrix.size();
 }
 
-const std::vector<int>& Graph::operator[](size_t index) const {
+const std::vector<size_t>& Graph::operator[](size_t index) const {
     return matrix[index];
 }
